syscall storage_obj_seek: fix sign extension
Fixes problem with sign extension (or lack thereof) for the syscall
storage_obj_seek. Updates the general rules of arguments for syscalls to
use signed 32-bit parameters when a signed parameter is needed.

Tested-by: Jerome Forissier <jerome.forissier@linaro.org> (HiKey)
Suggested-by: Jerome Forissier <jerome.forissier@linaro.org>
Tested-by: Jens Wiklander <jens.wiklander@linaro.org> (QEMU and FVP)

+++ core/include/tee/tee_svc_storage.h
@@ -75,7 +75,7 @@ TEE_Result syscall_storage_obj_write(unsigned long obj, void *data,
 
 TEE_Result syscall_storage_obj_trunc(unsigned long obj, size_t len);
 
-TEE_Result syscall_storage_obj_seek(unsigned long obj, long offset,
+TEE_Result syscall_storage_obj_seek(unsigned long obj, int32_t offset,
 				    unsigned long whence);
 
 void tee_svc_storage_close_all_enum(struct user_ta_ctx *utc);
